Migrate from haml to liquid
Haml is oooold news and the jekyll haml plugin has a dependency which
has a version mismatch with octopress.

Liquid can't run arbitrary code, so we need to do the date operations
with their tags, but that turns out to actually be fine when you get the
hang of it.