Fix SASS warnings
SASS is going to introduce a breaking change - I don't think it actually
affects this code at all, but the idea is that if you want rules to be
all applied in the same top level declarations, they need to be at the
top of the block. In this case show-links-on-hover needs to be below
that because it defines rules on hover.

  https://sass-lang.com/documentation/breaking-changes/mixed-decls/